+++ Raytracer by waev/Raytracer by waev/WUtilities.cpp
@@ -18,7 +18,7 @@ WColor getBackgroundCheckers(int x, int y, int width, int height)
 	int ratio = width / slices;
 	int sliceX = x / ratio;
 	int sliceY = y / ratio;
-	float sliceZ = sqrt(x*x + y*y);
+	float sliceZ = (float)sqrt(x*x + y*y);
 	float sliceVal = 1.0f / 6;
 	return WColor(sliceVal * sliceX, sliceVal * sliceY, 0.5);
 }
